HalkoMethods and HalkoClasses working
HalkoMethods can now be called from any class that derives HalkoClass. HalkoClasses add themselves to the halkoClasses-list of HalkoNetwork on Awake. When a halkoMethod is invoked the halkoClass will send a stream to the other clients, that contains "STREAM LENGTH (4 bytes) + STREAM FLAG (1 byte) +  HalkoClassIndex (4 bytes) + NUMBER OF PARAMETERS (4 bytes) + METHOD INDEX (4 bytes) + PARAMETERS CONVERTED TO BYTES
".
When receiving the stream, HalkoNetwork knows in which class the method is.